requirements.txt
torch
torchvision
matplotlib
pillow
gradio
